# update cli perf time limit