--- a/src/promptflow/promptflow/_cli/_pf_azure/entry.py
+++ b/src/promptflow/promptflow/_cli/_pf_azure/entry.py
@@ -104,6 +104,10 @@ def _get_workspace_info(args):
         return {}
 
 
+def _get_custom_dimensions(args):
+    return _get_workspace_info(args)
+
+
 def entry(argv):
     """
     Control plane CLI tools for promptflow cloud version.
@@ -112,7 +116,7 @@ def entry(argv):
     if hasattr(args, "user_agent"):
         setup_user_agent_to_operation_context(args.user_agent)
     logger = get_telemetry_logger()
-    custom_dimensions = _get_workspace_info(args)
+    custom_dimensions = _get_custom_dimensions(args)
     with log_activity(
         logger,
         _get_cli_activity_name(cli=prog, args=args),