--- a/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
@@ -1,3 +1,4 @@
+import os
 import sys
 import timeit
 from typing import Callable
@@ -5,6 +6,7 @@ from unittest import mock
 
 import pytest
 
+from promptflow._cli._pf_azure.entry import _get_custom_dimensions
 from promptflow._cli._user_agent import USER_AGENT as CLI_USER_AGENT  # noqa: E402
 from promptflow._sdk._utils import ClientUserAgentUtil
 from sdk_cli_azure_test.recording_utilities import is_replay
@@ -13,12 +15,20 @@ FLOWS_DIR = "./tests/test_configs/flows"
 DATAS_DIR = "./tests/test_configs/datas"
 
 
+def get_custom_dimensions(*args, **kwargs):
+    custom_dimensions = _get_custom_dimensions(*args, **kwargs)
+    custom_dimensions["pr_number"] = os.environ.get("PR_NUMBER")
+
+    return custom_dimensions
+
+
 def run_cli_command(cmd, time_limit=3600):
     from promptflow._cli._pf_azure.entry import main
 
     sys.argv = list(cmd)
     st = timeit.default_timer()
-    with mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun:
+    with (mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun,
+          mock.patch("promptflow._cli._pf_azure.entry._get_custom_dimensions", side_effect=get_custom_dimensions)):
         # Client side will modify user agent only through ClientUserAgentUtil to avoid impact executor/runtime.
         get_user_agent_fun.return_value = f"{CLI_USER_AGENT} perf_monitor/1.0"
         user_agent = ClientUserAgentUtil.get_user_agent()