--- a/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
@@ -17,7 +17,9 @@ DATAS_DIR = "./tests/test_configs/datas"
 
 def get_custom_dimensions(*args, **kwargs):
     custom_dimensions = _get_custom_dimensions(*args, **kwargs)
-    custom_dimensions["flow_run_id"] = os.environ.get("FLOW_RUN_ID")
+    custom_dimensions["flow run: "] = "https://github.com/microsoft/promptflow/actions/runs/{0}".format(
+        os.environ.get("FLOW_RUN_ID")
+    )
 
     return custom_dimensions