--- a/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
@@ -38,10 +38,9 @@ def run_cli_command(cmd, time_limit=3600):
 
     sys.argv = list(cmd)
     st = timeit.default_timer()
-    with (mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun,
-          mock.patch("promptflow._sdk._telemetry.activity.log_activity", side_effect=mock_log_activity),
-          mock.patch("promptflow._cli._pf_azure.entry.log_activity", side_effect=mock_log_activity),
-          ):
+    with mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun, mock.patch(
+        "promptflow._sdk._telemetry.activity.log_activity", side_effect=mock_log_activity
+    ), mock.patch("promptflow._cli._pf_azure.entry.log_activity", side_effect=mock_log_activity):
         # Client side will modify user agent only through ClientUserAgentUtil to avoid impact executor/runtime.
         get_user_agent_fun.return_value = f"{CLI_USER_AGENT} perf_monitor/1.0"
         user_agent = ClientUserAgentUtil.get_user_agent()