--- a/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
@@ -1,6 +1,7 @@
 import contextlib
 import io
 import multiprocessing
+import os
 import sys
 import tempfile
 import timeit
@@ -10,12 +11,22 @@ from unittest import mock
 
 import pytest
 
+from promptflow._cli._pf.entry import _get_custom_dimensions
 from promptflow._cli._user_agent import USER_AGENT as CLI_USER_AGENT  # noqa: E402
 from promptflow._sdk._utils import ClientUserAgentUtil
 
 FLOWS_DIR = "./tests/test_configs/flows"
 CONNECTIONS_DIR = "./tests/test_configs/connections"
 DATAS_DIR = "./tests/test_configs/datas"
+
+
+def get_custom_dimensions():
+    custom_dimensions = _get_custom_dimensions()
+    custom_dimensions["flow run: "] = "https://github.com/microsoft/promptflow/actions/runs/{0}".format(
+        os.environ.get("FLOW_RUN_ID")
+    )
+
+    return custom_dimensions
 
 
 def run_cli_command(cmd, time_limit=3600, result_queue=None):
@@ -25,9 +36,9 @@ def run_cli_command(cmd, time_limit=3600, result_queue=None):
     output = io.StringIO()
 
     st = timeit.default_timer()
-    with contextlib.redirect_stdout(output), mock.patch.object(
-        ClientUserAgentUtil, "get_user_agent"
-    ) as get_user_agent_fun:
+    with (contextlib.redirect_stdout(output),
+          mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun,
+          mock.patch("promptflow._cli._pf.entry._get_custom_dimensions", side_effect=get_custom_dimensions)):
         # Client side will modify user agent only through ClientUserAgentUtil to avoid impact executor/runtime.
         get_user_agent_fun.return_value = f"{CLI_USER_AGENT} perf_monitor/1.0"
         user_agent = ClientUserAgentUtil.get_user_agent()