--- a/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
@@ -41,7 +41,7 @@ def run_cli_command(cmd, time_limit=3600, result_queue=None):
     st = timeit.default_timer()
     with (contextlib.redirect_stdout(output),
           mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun,
-          mock.patch("promptflow._sdk._telemetry.log_activity", side_effect=mock_log_activity)):
+          mock.patch("promptflow._sdk._telemetry.activity.log_activity", side_effect=mock_log_activity)):
         # Client side will modify user agent only through ClientUserAgentUtil to avoid impact executor/runtime.
         get_user_agent_fun.return_value = f"{CLI_USER_AGENT} perf_monitor/1.0"
         user_agent = ClientUserAgentUtil.get_user_agent()