--- a/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
+++ b/src/promptflow/tests/sdk_cli_test/e2etests/test_cli_perf.py
@@ -24,7 +24,13 @@ def mock_log_activity(*args, **kwargs):
     custom_message = "flow run: https://github.com/microsoft/promptflow/actions/runs/{0}".format(
         os.environ.get("FLOW_RUN_ID")
     )
-    if "custom_dimensions" in kwargs and kwargs["custom_dimensions"] is not None:
+    if len(args) == 4:
+        if args[3] is not None:
+            args[3]["custom_message"] = custom_message
+        else:
+            args = list(args)
+            args[3] = {"custom_message": custom_message}
+    elif "custom_dimensions" in kwargs and kwargs["custom_dimensions"] is not None:
         kwargs["custom_dimensions"]["custom_message"] = custom_message
     else:
         kwargs["custom_dimensions"] = {"custom_message": custom_message}
@@ -41,7 +47,9 @@ def run_cli_command(cmd, time_limit=3600, result_queue=None):
     st = timeit.default_timer()
     with (contextlib.redirect_stdout(output),
           mock.patch.object(ClientUserAgentUtil, "get_user_agent") as get_user_agent_fun,
-          mock.patch("promptflow._sdk._telemetry.activity.log_activity", side_effect=mock_log_activity)):
+          mock.patch("promptflow._sdk._telemetry.activity.log_activity", side_effect=mock_log_activity),
+          mock.patch("promptflow._cli._pf.entry.log_activity", side_effect=mock_log_activity),
+          ):
         # Client side will modify user agent only through ClientUserAgentUtil to avoid impact executor/runtime.
         get_user_agent_fun.return_value = f"{CLI_USER_AGENT} perf_monitor/1.0"
         user_agent = ClientUserAgentUtil.get_user_agent()