# Description
After using replay mode

The time consumption of "**pf.azure.create**" related commands will be reduced, with an average time of 6 seconds, but the maximum time consumption is still 20 seconds+

The duration of other commands PC99 is less than 10 seconds, but the maximum duration is still 20 seconds+

This PR modifies the test mode to replay mode and reduces the time limit, and then continues to observe the average and maximum time consumption, PC99 time consumption.


# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated review from promptflow team. Learn more: [suggested workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an informative message. This means that previously merged commits do not appear in the history of the PR. For more information on cleaning up the commits in your PR, [see this page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.